vim: Adjust highlights to account for gitgutter behavior changes
`vim-gitgutter` formerly linked `SignColumn` to `LineNr`, but now it
doesn't anymore.  This exposed a flaw in that the background color
should be reset so that it matches the `Normal` highlight group's
background color.

With the new fall-back color behavior, the foreground colors don't need
to be explicitly set because the `Diff<Type>` highlight groups share the
same foreground color.  When `vim-gitgutter` detects this situation, it
ends up setting the exact same foreground color values.  The highlight
commands are left until I'm confident that the `Diff*` highlight groups
won't be changing in `dim-ansi`.