neovim: Port key mappings to Lua
A helper function is provided to simplify mapping creation.  The
`{opts}` are optional, and the key mappings are silent and
non-recursive by default.  Due to the default options being specified,
the port to Lua is not strictly one-to-one.  However, there is no
observable change in behavior because the previous mappings were silent
and non-recursive before.

The `<C-N>` key mapping to show the highlight groups under the cursor is
simplified because Neovim is known to provide the functions necessary.