neovim: Port auto commands to Lua
The new Lua automatic commands APIs allow for a Lua function to be
triggered in response to registered event.

As a workaround to `:help restore-cursor`, we register with `BufEnter`
instead for restoring the cursor because `&ft` will be set. This differs
from `vim` because Neovim enables Filetype detection by default, but it
occurs after `init.(vim|lua)`, which is why `&ft` is not available
during `BufRead(Post)`.