neovim: Use one shot BufRead workaround for restoring the cursor
When initially porting the Vimscript auto commands to Lua, I introduced
a bug that would move the cursor to the last opened position when
switching between windows. The root cause was using the `BufEnter` event
that was firing every time. I cheekily used `BufEnter` because the
buffer `filetype` was set when fired, whereas it is not set when
`BufRead` is called. In short, this is happening because Neovim
`:filetype plugin indent on` is being executed by default *after* the
user startup scripts.

To workaround this and desiring to keep everything in Lua, I ported a
Vimscript workaround provided upstream to Lua. It's definitely more
complicated and verbose than the Vimscript equivalent. However, it works
like a charm :).