tmux: Set TERM to tmux-256color, if possible; otherwise screen-256color
To support italics, either the tmux or tmux-256color TERM should be
used. Recent versions of ncurses provide the tmux-256color terminfo.
However, macOS (at the time of this writing) and systems with older
versions of ncurses won't have it. If it is not available, attempt to
bootstrap by installing tmux-256color ourselves to be used.