kitty: Add kitty terminal emulator
I discovered `kitty` terminal emulator recently. I like it enough to be
used as my default terminal emulator because it is configurable via a
text file and supports macOS and Linux — keeping in the spirit of using
tools that are as cross-platform as possible.

An official Nord port for kitty doesn't exist (i.e.,
https://github.com/nordtheme/nord/issues/104), and the Nord theme from
https://github.com/kovidgoyal/kitty-themes doesn't adhere to the color
palette recommendations. I'll use this until the official port exists to
move over to.