neovim: Update nvim-treesitter to 'main' branch
The 'main' branch of the nvim-treesitter plugin is a complete rewrite
and focuses purely on parser management. With the rewrite removing the
convenience of enabling the treesitter syntax highlighting, I am
removing the pre-installed parsers, as I need to revisit my overall
configuration Neovim configuration to ensure I am using treesitter
syntax highlighting for the languages  I care about.